Optimizing the draw command. Now only the faces that can be visible are drawn.

+++ src/world.h
@@ -11,5 +11,5 @@ typedef struct HxfWorld {
     /**
      * @brief An array containing all the cubes of the world.
      */
-    uint8_t cubes[HXF_WORLD_LENGTH][HXF_WORLD_LENGTH][HXF_WORLD_LENGTH];
+    uint32_t cubes[HXF_WORLD_LENGTH][HXF_WORLD_LENGTH][HXF_WORLD_LENGTH];
 } HxfWorld;
